# apispec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Golang RESTful API",
        "description": "API spec for golang RESTful API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api"
        }
    ],
    "paths": {
        "/cakes": {
            "get": {
                "tags": [
                    "Cake API"
                ],
                "security": [
                    {
                        "CakeAuth": []
                    }
                ],
                "summary": "List All Cakes",
                "description": "List All Cakes",
                "responses": {
                    "200": {
                        "description": "Success Get All Cakes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cake"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Cake API"
                ],
                "security": [
                    {
                        "CakeAuth": []
                    }
                ],
                "summary": "Create New Cake",
                "description": "Create New Cake",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateCake"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Create Cake",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Cake"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                }
            }
        },
        "/cakes/{cakeId}": {
            "get": {
                "tags": [
                    "Cake API"
                ],
                "security": [
                    {
                        "CakeAuth": []
                    }
                ],
                "summary": "Get Cake by Id",
                "description": "Get Cake by Id",
                "parameters": [
                    {
                        "name": "cakeId",
                        "in": "path",
                        "description": "Cake Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Find Cake",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cake"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cake API"
                ],
                "security": [
                    {
                        "CakeAuth": []
                    }
                ],
                "summary": "Update Cake by Id",
                "description": "Update Cake by Id",
                "parameters": [
                    {
                        "name": "cakeId",
                        "in": "path",
                        "description": "Cake Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateCake"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success Update Cake",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Cake"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cake API"
                ],
                "security": [
                    {
                        "CakeAuth": []
                    }
                ],
                "summary": "Delete Cake by Id",
                "description": "Delete Cake by Id",
                "parameters": [
                    {
                        "name": "cakeId",
                        "in": "path",
                        "description": "Cake Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Delete Cake",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Cake": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                }
            },
            "CreateOrUpdateCake": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "required": [
                    "title"
                ],
                "example": {
                    "title": "Lemon cheesecake",
                    "description": "A cheesecake made of lemon",
                    "rating": 7.3,
                    "image": "https://img.taste.com.au/ynYrqkOs/w720-h480-cfill-q80/taste/2016/11/sunny-lemon-cheesecake-102220-1.jpeg"
                }
            }
        },
        "responses": {
            "NotFound": {
                "description": "Error: Record not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "number"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "data": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "Unauthorized": {
                "description": "Error: UNAUTHORIZED",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data":{
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "CakeAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "Fill input with value : RAHASIA"
            }
        }
    }
}